- Implemented partial windows support - Added separate unit tests for windows   and make sure that the linux test don't run on a windows box   because they generate different things and will generate even   more different things in the future
Note:
I guess the *nix tests are totaly broken on a linux box.
This happens because we can't handle test dependencies anymore with eggs.
The changes in zc.buildout broke all the tests in this package
and nobody was recognizing it. I totaly dislike it that we do not have
a trunk centric development because this will make sure that we can run 
tests over at least a minimal set of core components. I'm sure that we
sometimes later will switch back to a trunk development. 

I really believe that development and distribution are two different 
things we mix together right now in our development process.

git-svn-id: file:///home/mg/src/zope-git-migration/zope-mirror/zc.zope3recipes/trunk@78476 62d5b8a3-27da-0310-9561-8e5933582275